refactor extension with rte support
